+++ pokemon-wiki-web/src/main/java/cn/fantasticmao/pokemon/web/controller/WeChatController.java
@@ -2,7 +2,8 @@ package cn.fantasticmao.pokemon.web.controller;
 
 import cn.fantasticmao.mundo.web.support.wechat.*;
 import cn.fantasticmao.pokemon.web.service.WechatService;
-import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,9 +19,9 @@ import javax.servlet.http.HttpServletRequest;
  * @author fantasticmao
  * @since 2018/12/5
  */
-@Slf4j
 @RestController
 public class WeChatController extends WeChatServerConfig {
+    private static final Logger LOGGER = LoggerFactory.getLogger(WeChatController.class);
     @Resource
     private WechatService wechatService;
 
@@ -32,17 +33,17 @@ public class WeChatController extends WeChatServerConfig {
 
     @PostMapping(value = "/wechat", produces = MediaType.TEXT_XML_VALUE)
     public String message(HttpServletRequest request, @RequestBody String xml) {
-        log.debug("接收微信消息 xml={}", xml);
+        LOGGER.debug("接收微信消息 xml={}", xml);
         WeChatMessage weChatMessage = WeChatMessageFactory.newMessage(xml);
         if (weChatMessage.getMsgType().equals(WeChatMessageType.TEXT)) {
             WeChatTextMessage weChatTextMessage = WeChatMessageFactory.newTextMessage(xml);
             final String content = weChatTextMessage.getContent();
-            log.debug("接收用户发送的普通消息 content={}", content);
+            LOGGER.debug("接收用户发送的普通消息 content={}", content);
 
             final String replyContent = wechatService.searchPokemonInfosByName(content);
             final String replyContentXml = WeChatMessageFactory.newXmlByTextMessage(
                 weChatTextMessage.getFromUserName(), weChatTextMessage.getToUserName(), replyContent);
-            log.debug("被动回复微信消息 xml={}", replyContentXml);
+            LOGGER.debug("被动回复微信消息 xml={}", replyContentXml);
             return replyContentXml;
         } else {
             return "success";