+++ pokemon-wiki-web/src/test/java/cn/fantasticmao/pokemon/web/SpringTest.java
@@ -1,13 +1,8 @@
 package cn.fantasticmao.pokemon.web;
 
 import cn.fantasticmao.pokemon.Application;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ApplicationContext;
-
-import javax.annotation.Resource;
 
 /**
  * SpringTest
@@ -18,11 +13,4 @@ import javax.annotation.Resource;
 @AutoConfigureMockMvc
 @SpringBootTest(classes = Application.class)
 public class SpringTest {
-    @Resource
-    private ApplicationContext applicationContext;
-
-    @Test
-    public void testSpring() {
-        Assertions.assertNotNull(applicationContext);
-    }
 }