+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2022-05-12
+
+- 升级 `cn.fantasticmao.mundo:mundo-all:1.0.5`
+- 清理代码
+
 ## 2022-03-03
 
 - 升级 JDK 版本至 11